Separate hitrecord to own file. Fix tests.
